+++ lib/src/tinycolor.dart
@@ -1,11 +1,10 @@
 import 'dart:math' as math;
 import 'dart:ui' show Color;
 
-import 'package:flutter/painting.dart' show HSVColor;
+import 'package:flutter/painting.dart' show HSLColor, HSVColor;
 import 'package:pigment/pigment.dart';
 
 import 'conversion.dart';
-import 'hsl_color.dart';
 import 'util.dart';
 
 class TinyColor {
@@ -17,15 +16,16 @@ class TinyColor {
 
   TinyColor.fromColor(this.originalColor) : _color = Color(originalColor.value);
 
+  @Deprecated('Use TinyColor(Color.fromARGB(a, r, g, b)) instead.')
   factory TinyColor.fromRGB({
     required int r,
     required int g,
     required int b,
-    int a = 100,
+    int a = 255,
   }) =>
       TinyColor.fromColor(Color.fromARGB(a, r, g, b));
 
-  factory TinyColor.fromHSL(HslColor hsl) => TinyColor.fromColor(hslToColor(hsl));
+  factory TinyColor.fromHSL(HSLColor hsl) => TinyColor(hsl.toColor());
 
   @Deprecated('Use TinyColor.fromHSVColor() instead.')
   factory TinyColor.fromHSV(HSVColor hsl) => TinyColor.fromHSVColor(hsl);
@@ -59,19 +59,7 @@ class TinyColor {
 
   HSVColor toHSVColor() => HSVColor.fromColor(_color);
 
-  HslColor toHsl() {
-    final hsl = rgbToHsl(
-      r: _color.red.toDouble(),
-      g: _color.green.toDouble(),
-      b: _color.blue.toDouble(),
-    );
-    return HslColor(
-      h: hsl.h * 360,
-      s: hsl.s,
-      l: hsl.l,
-      a: _color.alpha.toDouble(),
-    );
-  }
+  HSLColor toHsl() => HSLColor.fromColor(_color);
 
   String toHex8() => _color.value.toRadixString(16).padLeft(8, '0');
 
@@ -79,9 +67,8 @@ class TinyColor {
 
   TinyColor lighten([int amount = 10]) {
     final hsl = toHsl();
-    hsl.l += amount / 100;
-    hsl.l = clamp01(hsl.l);
-    return TinyColor.fromHSL(hsl);
+    final lightness = clamp01(hsl.lightness + amount / 100);
+    return TinyColor.fromHSL(hsl.withLightness(lightness));
   }
 
   TinyColor brighten([int amount = 10]) {
@@ -114,62 +101,55 @@ class TinyColor {
 
   TinyColor darken([int amount = 10]) {
     final hsl = toHsl();
-    hsl.l -= amount / 100;
-    hsl.l = clamp01(hsl.l);
-    return TinyColor.fromHSL(hsl);
+    final lightness = clamp01(hsl.lightness - amount / 100);
+    return TinyColor.fromHSL(hsl.withLightness(lightness));
   }
 
   TinyColor tint([int amount = 10]) => mix(
-        input: const Color.fromRGBO(255, 255, 255, 1.0),
-        amount: amount,
+        const Color.fromRGBO(255, 255, 255, 1.0),
+        amount,
       );
 
   TinyColor shade([int amount = 10]) => mix(
-        input: const Color.fromRGBO(0, 0, 0, 1.0),
-        amount: amount,
+        const Color.fromRGBO(0, 0, 0, 1.0),
+        amount,
       );
 
   TinyColor desaturate([int amount = 10]) {
     final hsl = toHsl();
-    hsl.s -= amount / 100;
-    hsl.s = clamp01(hsl.s);
-    return TinyColor.fromHSL(hsl);
+    final saturation = clamp01(hsl.saturation - amount / 100);
+    return TinyColor.fromHSL(hsl.withSaturation(saturation));
   }
 
   TinyColor saturate([int amount = 10]) {
     final hsl = toHsl();
-    hsl.s += amount / 100;
-    hsl.s = clamp01(hsl.s);
-    return TinyColor.fromHSL(hsl);
+    final saturation = clamp01(hsl.saturation + amount / 100);
+    return TinyColor.fromHSL(hsl.withSaturation(saturation));
   }
 
   TinyColor greyscale() => desaturate(100);
 
   TinyColor spin(double amount) {
     final hsl = toHsl();
-    final hue = (hsl.h + amount) % 360;
-    hsl.h = hue < 0 ? 360 + hue : hue;
-    return TinyColor.fromHSL(hsl);
+    final hue = (hsl.hue + amount) % 360;
+    return TinyColor.fromHSL(hsl.withHue(hue < 0 ? 360 + hue : hue));
   }
 
-  TinyColor mix({
-    required Color input,
-    int amount = 50,
-  }) {
+  TinyColor mix(Color toColor, [int amount = 50]) {
     final p = amount / 100.0;
     final color = Color.fromARGB(
-      ((input.alpha - _color.alpha) * p + _color.alpha).round(),
-      ((input.red - _color.red) * p + _color.red).round(),
-      ((input.green - _color.green) * p + _color.green).round(),
-      ((input.blue - _color.blue) * p + _color.blue).round(),
+      ((toColor.alpha - _color.alpha) * p + _color.alpha).round(),
+      ((toColor.red - _color.red) * p + _color.red).round(),
+      ((toColor.green - _color.green) * p + _color.green).round(),
+      ((toColor.blue - _color.blue) * p + _color.blue).round(),
     );
     return TinyColor.fromColor(color);
   }
 
   TinyColor complement() {
     final hsl = toHsl();
-    hsl.h = (hsl.h + 180) % 360;
-    return TinyColor.fromHSL(hsl);
+    final hue = (hsl.hue + 180) % 360;
+    return TinyColor.fromHSL(hsl.withHue(hue));
   }
 
   Color get color => _color;